refactor: httpResponse 패키지 생성
