+++ src/main/java/webserver/httpResponse/HttpResponse.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.httpResponse;
 
 import java.io.DataOutputStream;
 import java.io.IOException;
@@ -8,6 +8,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import utils.ContentType;
+import webserver.View;
 
 public class HttpResponse {
 	private static final Logger logger = LoggerFactory.getLogger(HttpResponse.class);
